+++ src/main/java/in/koreatech/payment/repository/TemporaryPaymentRepository.java
@@ -1,0 +1,10 @@
+package in.koreatech.payment.repository;
+
+import org.springframework.data.repository.Repository;
+
+import in.koreatech.payment.model.TemporaryPayment;
+
+public interface TemporaryPaymentRepository extends Repository<TemporaryPayment, Integer> {
+
+    void save(TemporaryPayment temporaryPayment);
+}
