+++ src/main/java/in/koreatech/payment/service/TossService.java
@@ -3,8 +3,8 @@ package in.koreatech.payment.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import in.koreatech.payment.model.redis.TemporaryPayment;
-import in.koreatech.payment.repository.TemporaryPaymentRedisRepository;
+import in.koreatech.payment.model.TemporaryPayment;
+import in.koreatech.payment.repository.TemporaryPaymentRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -12,11 +12,11 @@ import lombok.RequiredArgsConstructor;
 @Transactional(readOnly = true)
 public class TossService implements PaymentService {
 
-    private final TemporaryPaymentRedisRepository temporaryPaymentRedisRepository;
+    private final TemporaryPaymentRepository temporaryPaymentRepository;
 
     @Transactional
     public void saveTemporaryPayment(String orderId, Integer userId, Integer amount) {
         TemporaryPayment temporaryPayment = TemporaryPayment.of(orderId, userId, amount);
-        temporaryPaymentRedisRepository.save(temporaryPayment);
+        temporaryPaymentRepository.save(temporaryPayment);
     }
 }
